[v6.x] combine seedfinders-mod into racemod
